feat: Add dashboard UI with capacity overview, detailed shelter list, and statistics, supported by new data types and utility functions.
