+++ src/app/dashboard/page.tsx
@@ -2,27 +2,10 @@
 import { useState, useEffect, useCallback } from 'react';
 import axios from 'axios';
 import ExcelJS from 'exceljs';
-
-interface Shelter {
-  _id: string;
-  name: string;
-  district: string;
-  subdistrict?: string;
-  capacity: number;
-  currentOccupancy: number;
-  capacityStatus?: string;
-  dailyLogs?: { date: string; checkIn: number; checkOut: number }[];
-  updatedAt?: string;
-}
-
-interface Stats {
-  totalShelters: number;
-  totalCapacity: number;
-  totalOccupancy: number;
-  criticalShelters: number;
-  warningShelters: number;
-  totalMedicalRequests: number;
-}
+import { Shelter, Stats } from "@/types/shelter";
+import StatsGrid from '@/components/dashboard/StatsGrid';
+import CapacityOverview from '@/components/dashboard/CapacityOverview';
+import ShelterList from '@/components/dashboard/ShelterList';
 
 export default function UnifiedDashboard() {
   const [shelters, setShelters] = useState<Shelter[]>([]);
@@ -30,9 +13,7 @@ export default function UnifiedDashboard() {
   const [searchTerm, setSearchTerm] = useState('');
   const [loading, setLoading] = useState(true);
   const [isExporting, setIsExporting] = useState(false);
-  const [currentPage, setCurrentPage] = useState(1);
   const [timeRange, setTimeRange] = useState(1); // Default to 1 day (Today)
-  const ITEMS_PER_PAGE = 30; // แสดงแถวได้มากขึ้นในรูปแบบตารางเดียวพอดีสวยๆ
 
   const fetchData = useCallback(async () => {
     try {
@@ -101,48 +82,6 @@ export default function UnifiedDashboard() {
     }
   };
 
-  const filteredShelters = shelters.filter(s => 
-    (s.name?.toLowerCase().includes(searchTerm.toLowerCase())) || 
-    (s.district?.toLowerCase().includes(searchTerm.toLowerCase()))
-  );
-
-  // Helper เพื่อคำนวณยอดสะสมตามช่วงเวลา
-  const getAggregatedMovement = (logs: Shelter['dailyLogs']) => {
-    if (!logs || !Array.isArray(logs)) return { in: 0, out: 0 };
-    
-    // สร้าง Array ของวันที่ย้อนหลังตาม timeRange (ยึดตามเขตเวลาประเทศไทย UTC+7) รูปแบบ YYYY-MM-DD
-    const targetDates: string[] = [];
-    const now = new Date();
-    
-    for (let i = 0; i < timeRange; i++) {
-      const d = new Date(now.getTime() + (7 * 60 * 60 * 1000));
-      d.setDate(d.getDate() - i);
-      const dateStr = d.toISOString().split('T')[0];
-      targetDates.push(dateStr);
-    }
-
-    return logs
-      .filter(log => targetDates.includes(log.date))
-      .reduce((acc, log) => ({
-        in: acc.in + (Number(log.checkIn) || 0),
-        out: acc.out + (Number(log.checkOut) || 0)
-      }), { in: 0, out: 0 });
-  };
-
-  // Pagination Logic
-  const totalPages = Math.ceil(filteredShelters.length / ITEMS_PER_PAGE);
-  const paginatedShelters = filteredShelters.slice(
-    (currentPage - 1) * ITEMS_PER_PAGE,
-    currentPage * ITEMS_PER_PAGE
-  );
-
-  // Reset page when search changes
-  useEffect(() => {
-    setCurrentPage(1);
-  }, [searchTerm]);
-
-  const occupancyRate = stats ? (stats.totalOccupancy / (stats.totalCapacity || 1)) * 100 : 0;
-
   if (loading) return (
     <div className="container py-5 text-center">
       <div className="spinner-border text-primary" role="status"></div>
@@ -168,358 +107,20 @@ export default function UnifiedDashboard() {
         </div>
       </div>
 
-      {/* ส่วนที่ 2: การ์ดตัวเลขสรุป (จากหน้ารายงาน) */}
-      {stats && (
-        <div className="row g-3 mb-4">
-          <div className="col-12 col-sm-6 col-lg-3">
-            <div className="card h-100 border-0 shadow-sm overflow-hidden text-white" style={{ background: 'linear-gradient(45deg, #0d6efd, #0043a8)', borderRadius: '15px' }}>
-              <div className="card-body p-3 p-md-4 position-relative">
-                <i className="bi bi-house-door-fill position-absolute bottom-0 end-0 opacity-25 me-3 mb-2" style={{ fontSize: '2rem' }}></i>
-                <div className="text-white fw-bold mb-1" style={{ fontSize: '0.8rem', letterSpacing: '0.5px' }}>จำนวนศูนย์ทั้งหมด</div>
-                <h2 className="mb-0 fw-bold" style={{ fontSize: '1.8rem' }}>{stats.totalShelters.toLocaleString()} <small className="fs-6 fw-normal opacity-75">แห่ง</small></h2>
-              </div>
-            </div>
-          </div>
-          <div className="col-12 col-sm-6 col-lg-3">
-            <div className="card h-100 border-0 shadow-sm overflow-hidden text-white" style={{ background: 'linear-gradient(45deg, #dc3545, #a71d2a)', borderRadius: '15px' }}>
-              <div className="card-body p-3 p-md-4 position-relative">
-                <i className="bi bi-exclamation-triangle-fill position-absolute bottom-0 end-0 opacity-25 me-3 mb-2" style={{ fontSize: '2rem' }}></i>
-                <div className="text-white fw-bold mb-1" style={{ fontSize: '0.8rem', letterSpacing: '0.5px' }}>ศูนย์ในสถานะ &quot;ล้น&quot;</div>
-                <h2 className="mb-0 fw-bold" style={{ fontSize: '1.8rem' }}>{stats.criticalShelters} <small className="fs-6 fw-normal opacity-75">แห่ง</small></h2>
-              </div>
-            </div>
-          </div>
-          <div className="col-12 col-sm-6 col-lg-3">
-            <div className="card h-100 border-0 shadow-sm overflow-hidden" style={{ background: 'linear-gradient(45deg, #ffc107, #ff8f00)', color: '#212529', borderRadius: '15px' }}>
-              <div className="card-body p-3 p-md-4 position-relative">
-                <i className="bi bi-house-exclamation-fill position-absolute bottom-0 end-0 opacity-25 me-3 mb-2" style={{ fontSize: '2rem' }}></i>
-                <div className="fw-bold mb-1" style={{ fontSize: '0.8rem', letterSpacing: '0.5px' }}>ศูนย์ที่ &quot;ใกล้เต็ม&quot;</div>
-                <h2 className="mb-0 fw-bold" style={{ fontSize: '1.8rem' }}>{stats.warningShelters} <small className="fs-6 fw-normal opacity-75">แห่ง</small></h2>
-              </div>
-            </div>
-          </div>
-          <div className="col-12 col-sm-6 col-lg-3">
-            <div className="card h-100 border-0 shadow-sm overflow-hidden text-white" style={{ background: 'linear-gradient(45deg, #0dcaf0, #00acc1)', borderRadius: '15px' }}>
-              <div className="card-body p-3 p-md-4 position-relative">
-                <i className="bi bi-capsule-pill position-absolute bottom-0 end-0 opacity-25 me-3 mb-2" style={{ fontSize: '2rem' }}></i>
-                <div className="text-white fw-bold mb-1" style={{ fontSize: '0.8rem', letterSpacing: '0.5px' }}>คำร้องขอยาทั้งหมด</div>
-                <h2 className="mb-0 fw-bold" style={{ fontSize: '1.8rem' }}>{stats.totalMedicalRequests} <small className="fs-6 fw-normal opacity-75">รายการ</small></h2>
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
+      {/* ส่วนที่ 2: การ์ดตัวเลขสรุป */}
+      {stats && <StatsGrid stats={stats} />}
 
       {/* ส่วนที่ 3: แถบความหนาแน่นรวม */}
-      <div className="card shadow-sm border-theme mb-4">
-        <div className="card-body py-3">
-          <div className="d-flex justify-content-between align-items-center mb-2">
-            <h6 className="mb-0 fw-bold" style={{ color: 'var(--text-primary)' }}>ระดับความหนาแน่นผู้อพยพภาพรวมทั้งจังหวัด</h6>
-            <span className="badge bg-light text-dark">{stats?.totalOccupancy.toLocaleString()} / {stats?.totalCapacity.toLocaleString()} คน ({occupancyRate.toFixed(1)}%)</span>
-          </div>
-          <div className="progress" style={{ height: '12px', borderRadius: '6px' }}>
-            <div 
-              className={`progress-bar progress-bar-striped progress-bar-animated ${occupancyRate > 90 ? 'bg-danger' : occupancyRate > 75 ? 'bg-warning' : 'bg-success'}`}
-              style={{ width: `${Math.min(occupancyRate, 100)}%` }}
-            ></div>
-          </div>
-        </div>
-      </div>
+      {stats && <CapacityOverview stats={stats} />}
 
       {/* ส่วนที่ 4: รายการศูนย์พักพิงแบบตาราง */}
-      <div className="card shadow-sm border-0 mb-3" style={{ backgroundColor: 'var(--bg-card)' }}>
-        <div className="card-header bg-transparent border-bottom py-3">
-          <div className="row g-3 align-items-center">
-            <div className="col-12 col-xl-4 text-center text-xl-start">
-              <h5 className="mb-0 fw-bold" style={{ color: 'var(--text-primary)' }}>
-                📍 ความเคลื่อนไหวรายศูนย์ {timeRange === 1 ? '(วันนี้)' : `(ย้อนหลัง ${timeRange} วัน)`}
-              </h5>
-            </div>
-            <div className="col-12 col-md-7 col-xl-4 d-flex justify-content-center">
-              <div className="btn-group btn-group-sm p-1 rounded-pill overflow-auto w-100 w-md-auto" style={{ backgroundColor: 'rgba(0,0,0,0.05)', whiteSpace: 'nowrap' }}>
-                {[1, 3, 7, 14, 30].map((range) => (
-                  <button 
-                    key={range}
-                    className={`btn px-3 rounded-pill border-0 ${timeRange === range ? 'btn-primary shadow-sm' : 'text-secondary'}`}
-                    onClick={() => setTimeRange(range)}
-                  >
-                    {range === 1 ? 'วันนี้' : `${range} วัน`}
-                  </button>
-                ))}
-              </div>
-            </div>
-            <div className="col-12 col-md-5 col-xl-4">
-              <div className="position-relative">
-                <i className="bi bi-search position-absolute top-50 start-0 translate-middle-y ms-3 text-secondary"></i>
-                <input 
-                  type="text" 
-                  className="form-control form-control-sm ps-5 border-theme" 
-                  placeholder="ค้นหาชื่อศูนย์/สถานที่..." 
-                  value={searchTerm}
-                  onChange={(e) => setSearchTerm(e.target.value)}
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-        
-        <div className="table-responsive">
-          <table className="table table-hover align-middle mb-0 text-theme" style={{ fontSize: '0.9rem' }}>
-            <thead>
-              <tr className="small fw-bold opacity-75">
-                <th className="ps-4 py-3" style={{ width: '35%' }}>ชื่อศูนย์ / สถานที่</th>
-                <th className="py-3 d-none d-lg-table-cell">ตำบล / อำเภอ</th>
-                {timeRange > 0 && <th className="text-center py-3">ความเคลื่อนไหว</th>}
-                <th className="text-center py-3">ครองเตียง (%)</th>
-                <th className="py-3 d-none d-md-table-cell">ความจุรวม</th>
-                <th className="py-3">สถานะ</th>
-                <th className="pe-4 py-3 d-none d-xl-table-cell">อัปเดตล่าสุด</th>
-              </tr>
-            </thead>
-            <tbody>
-              {paginatedShelters.map((shelter) => {
-                const percent = (shelter.currentOccupancy / (shelter.capacity || 1)) * 100;
-                let statusColor = "success";
-                let statusText = "รองรับได้";
-                
-                if (percent >= 100) { statusColor = "danger"; statusText = "ล้นศูนย์"; }
-                else if (percent >= 80) { statusColor = "warning"; statusText = "ใกล้เต็ม"; }
-
-                const movement = getAggregatedMovement(shelter.dailyLogs);
-
-                return (
-                  <tr key={shelter._id} className="border-bottom-theme">
-                    <td className="ps-4 py-3">
-                      <div className="d-flex align-items-start gap-2">
-                        <i className="bi bi-geo-alt-fill text-danger mt-1 d-none d-sm-block"></i>
-                        <div>
-                          <div className="fw-bold mb-0 text-primary-theme" style={{ fontSize: '0.95rem' }}>{shelter.name}</div>
-                          <div className="small text-secondary d-lg-none">
-                            {shelter.subdistrict} {shelter.district}
-                          </div>
-                        </div>
-                      </div>
-                    </td>
-                    <td className="d-none d-lg-table-cell py-3">
-                      <div>{shelter.subdistrict}</div>
-                      <small className="text-secondary">{shelter.district}</small>
-                    </td>
-                    <td>
-                      <div className="d-flex justify-content-center gap-3">
-                        <span className="text-success fw-bold">+{movement.in}</span>
-                        <span className="text-danger fw-bold">-{movement.out}</span>
-                      </div>
-                    </td>
-                    <td className="text-center">
-                      <div style={{ width: '80px', margin: '0 auto' }}>
-                        <div className="d-flex justify-content-between x-small mb-1">
-                          <span className={`text-${statusColor}`}>{percent.toFixed(0)}%</span>
-                        </div>
-                        <div className="progress" style={{ height: '4px' }}>
-                          <div className={`progress-bar bg-${statusColor}`} style={{ width: `${Math.min(percent, 100)}%` }}></div>
-                        </div>
-                      </div>
-                    </td>
-                    <td className="text-center fw-bold d-none d-md-table-cell py-3">
-                       {shelter.currentOccupancy} / {shelter.capacity}
-                    </td>
-                    <td className="py-3">
-                      <span className={`badge rounded-pill bg-${statusColor}-subtle text-${statusColor} px-2 py-1`} style={{ fontSize: '0.75rem' }}>
-                        {statusText}
-                      </span>
-                    </td>
-                    <td className="pe-4 text-secondary small d-none d-xl-table-cell py-3">
-                      {new Date(shelter.updatedAt || Date.now()).toLocaleDateString('th-TH')}
-                    </td>
-                  </tr>
-                );
-              })}
-            </tbody>
-          </table>
-        </div>
-      </div>
-
-      {/* Pagination UI */}
-      {totalPages > 1 && (
-        <div className="d-flex justify-content-center mt-5 mb-5">
-          <nav className="custom-pagination">
-            <div className="pagination-container d-flex align-items-center">
-              {/* First Page */}
-              <button 
-                className="pag-btn" 
-                onClick={() => setCurrentPage(1)} 
-                disabled={currentPage === 1}
-              >
-                «
-              </button>
-              
-              {/* Prev Page */}
-              <button 
-                className="pag-btn" 
-                onClick={() => setCurrentPage(p => Math.max(1, p - 1))}
-                disabled={currentPage === 1}
-              >
-                ‹
-              </button>
-
-              {/* Page Numbers */}
-              {(() => {
-                const pages = [];
-                const showRange = 2; // จำนวนหน้าที่จะโชว์รอบๆ หน้าปัจจุบัน
-                
-                for (let i = 1; i <= totalPages; i++) {
-                  if (
-                    i === 1 || 
-                    i === totalPages || 
-                    (i >= currentPage - showRange && i <= currentPage + showRange)
-                  ) {
-                    pages.push(
-                      <button 
-                        key={i}
-                        className={`pag-btn ${currentPage === i ? 'active' : ''}`}
-                        onClick={() => setCurrentPage(i)}
-                      >
-                        {i}
-                      </button>
-                    );
-                  } else if (
-                    i === currentPage - showRange - 1 || 
-                    i === currentPage + showRange + 1
-                  ) {
-                    pages.push(<span key={i} className="pag-ellipsis">...</span>);
-                  }
-                }
-                return pages;
-              })()}
-
-              {/* Next Page */}
-              <button 
-                className="pag-btn" 
-                onClick={() => setCurrentPage(p => Math.min(totalPages, p + 1))}
-                disabled={currentPage === totalPages}
-              >
-                ›
-              </button>
-
-              {/* Last Page */}
-              <button 
-                className="pag-btn" 
-                onClick={() => setCurrentPage(totalPages)}
-                disabled={currentPage === totalPages}
-              >
-                »
-              </button>
-            </div>
-          </nav>
-        </div>
-      )}
-
-      <style jsx>{`
-        .custom-pagination .pagination-container {
-          background-color: var(--bg-secondary);
-          border: 1px solid var(--border-color);
-          border-radius: 50px;
-          overflow: hidden;
-          padding: 4px;
-          box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-        }
-        .pag-btn {
-          background: transparent;
-          border: none;
-          color: var(--text-secondary);
-          padding: 6px 14px;
-          min-width: 40px;
-          height: 38px;
-          border-radius: 50px;
-          transition: all 0.2s;
-          font-size: 0.9rem;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-        }
-        .pag-btn:hover:not(:disabled):not(.active) {
-          background-color: var(--table-hover);
-          color: var(--text-primary);
-        }
-        .pag-btn.active {
-          background-color: #0d6efd;
-          color: white;
-          font-weight: bold;
-          box-shadow: 0 4px 12px rgba(13, 110, 253, 0.4);
-        }
-        .pag-btn:disabled {
-          color: var(--text-secondary);
-          opacity: 0.4;
-          cursor: not-allowed;
-        }
-        .pag-ellipsis {
-          padding: 8px 12px;
-          color: var(--text-secondary);
-          background-color: transparent;
-          font-size: 0.8rem;
-          display: flex;
-          align-items: flex-end;
-          height: 100%;
-        }
-        .border-bottom-theme {
-          border-bottom: 1px solid var(--border-color);
-        }
-        .text-primary-theme {
-          color: var(--text-primary);
-        }
-        .bg-danger-subtle { background-color: rgba(220, 53, 69, 0.1) !important; }
-        .bg-warning-subtle { background-color: rgba(255, 193, 7, 0.1) !important; }
-        .bg-success-subtle { background-color: rgba(25, 135, 84, 0.1) !important; }
-        
-        .table-hover tbody tr:hover {
-          background-color: var(--table-hover);
-        }
-
-        /* Responsive Pagination Pills */
-        .page-btn-pill {
-          background: transparent;
-          color: var(--text-secondary);
-          border: none;
-          padding: 6px 12px;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          transition: all 0.2s;
-          cursor: pointer;
-        }
-        .page-btn-pill:disabled {
-          opacity: 0.3;
-          cursor: default;
-        }
-        .page-btn-pill:not(:disabled):hover {
-          color: #fff;
-        }
-        .page-num-pill {
-          background: transparent;
-          color: var(--text-secondary);
-          border: none;
-          min-width: 32px;
-          height: 32px;
-          border-radius: 50%;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          font-size: 0.85rem;
-          transition: all 0.2s;
-          cursor: pointer;
-        }
-        .page-num-pill:hover {
-          background: rgba(255, 255, 255, 0.05);
-          color: #fff;
-        }
-        .page-num-pill.active {
-          background: #0d6efd;
-          color: #fff;
-          font-weight: bold;
-          box-shadow: 0 4px 10px rgba(13, 110, 253, 0.3);
-        }
-      `}</style>
+      <ShelterList 
+        shelters={shelters}
+        timeRange={timeRange}
+        setTimeRange={setTimeRange}
+        searchTerm={searchTerm}
+        setSearchTerm={setSearchTerm}
+      />
     </div>
   );
 }
-
